Add ObjectInspector::TemplatingFormatter as default
Rename ObjectInspector::DefaultFormatter to
ObjectInspector::CombiningFormatter

Add ObjectInspector::TemplatingFormatter make it the new
default Formatter because it is ~30% faster.

Update the `flags` and `info` demarcation symbols in the
Formatters.

Stop using `autoload` in lib/object_inspector.rb since Matz
highly discourages it and plans to remove it in Ruby v3.