Add a basic GitHub Action for CI
This replaces travis-ci, since travis-ci is no longer a thing I can use.

Establishes a GitHub action that will run tests on Ruby versions
2.7, 3.0, 3.1, and 3.2.

Currently: Can't get version prior to 2.7 to run due to an
incompatibility in gem versions.

Also, remove the CodeClimate badge for Test Coverage since it requires
a CI service to upload that data to CodeClimate and I don't yet know
how to get GetHub Actions to do this.