Simplify template lookup in ObjectInspector::TemplatingFormatter
I had originally used separate class methods for each template name/type
because that would make it easier to override just one of them, but...
they're really just data so I'm now wanting a data object (Hash) to
house them. Plus, with Ruby 3.3, many memoizations can lead to slower
performance, so I'm now wanting to reduce that... so a single class
methood is a good balance between flexibility, speed, and simplicity of
concept.

This also fixes the documentation problem... I've now added 
documentation for this one method, which removes the burden of adding
documentation for the many different methods that were present before.
And now `yard` is reporting 100% documentation across the project.