Remove use of OpenStruct + Fix up test suite
We no longer use OpenStruct as it is slow and the library (ostrcut) is
being removed from Ruby core in v3.5

Also, add `issues` to the CombiningFormatter. This doesn't really effect
anything/anyone since this formatter was not being used and, being
slower, there's no reason to believe anyone was opting into using it
over the TemplatingFormatter. But... it was a incongruity in the test
suite so I wanted to clean it up.

Also, fix up the test suite:
- Remove use of OpenStruct, which was plentiful
- Refactor `klazz` -> `unit_class`, as has been my new standard for a
  while now
- Add missing tests for inclusion of `issues` in output Strings