Add ObjectInspector::InspectorsHelper#inspect!
Add ObjectInspector::InspectorsHelper#inspect! is meant to be used as a
short, quick version of
ObjectInspector::InspectorsHelper#inspect(scope: :all) -- basically just
passing in `scope: :all`.

This is helpful because it's not always easy to remember what to type
out when you just want to see all the details of an object. Which is
why `:all` does -- it's the "Wild Card" scope (will match on all
`scope?` queries). And the bang (!) version of the method makes sense.
It's "more dangerous" in the sense that it might perform extra queries
vs the non-bang version.