Don't wrap test files in class-level describe blocks
At some point in the past, class-level describe blocks stopped being
necessary... (not sure when). But, it used to be that if you didn't add
a class-level describe block, test errors would be context-less.

That is, where now we'll see e.g.:

```
 FAIL ObjectInspector::ScopeTest::#==#test_0002_returns false, GIVEN a Scope with a different #name (0.01s)
```

... previously we would have only seen:

```
 FAIL #==#test_0002_returns false, GIVEN a Scope with a different #name (0.01s)
```

^ Completely devoid of context as to what Class has failed!

Anyway, now that we see the test name in the failure message, we can get
rid of the top-most level describe blocks.

---

Also, move all test doubles (classes) to the bottom of test files. There
was a mix of organization strategies, so this makes them all the same.
It's pretty arbitrary as to where they live (top vs bottom).