Simplify usage option in README
Remove some of the more obscure or more manual usage options shown in
the README. e.g. it's be rare not to want to just use a the
InspectBehaviors mix-in, so don't describe how to do this (especially
not up-front).

Simplify the description of Scope Names as well. And move the part on
how to temporarily disable Object Inspector down the page.

The goal is to make it easier for new users to understand the core
behaviors of the gem.